(#11) Modem auto-connects during boot
Created a systemd service to auto connect the modem to the network.

This is implemented in a bash script that powers up the modem and, once
available, brings up the network interface.

It depends on the existing modem.py script.

NOTE: The modem.py script differs from the one that was tested with.