Disable installation of external linux-image.
Disable installation of linux-image located in an external directory.

It is intended that a linux-image debian package will be placed within
the directory structure of this repository and installed from there.

This will allow this project to be checked out in isolation and be able to
build without any other project dependencies.  It will also makes the
installaiton scripts much easier to maintain :)