IND-EFD: [#42] Fix KeyError in efd_app.py
efd_app.py was generating a key error due to a race condition accessing
a key in the app_state object.  This was fixed by using the dict get()
method and supplying a default value if key is not present.

Also provided math_isfinite() function to implement the math.isfinite()
function, which is only available in Python >= 3.2.

Updated version number to v0.4.1.