IND-EFD: Add watchdog deamon package.
Add the watchdog package.

NOTE: this is not being used at all as we are using systemd to manage the
hardware watchdog.  The watchdog daemon may come in handy for monitoring other
system functionality such as internet connectivity, memory usage, etc.

The watchdog does not startup at boot by default.
There is a hook for the watchdog to enalbe it but it is currently disabled.
We must be careful to ensure that only one daemon is manageing the hardware
watchdog (either systemd or watchdog).