IND-EFD: script to check internet and restart.
Add a script to check that the internet is up and a connection to the outside
world can be established by way of a ping to a well known reliable server.
If the ping fails for whatever reason the call the `modem-power-enable.sh`
to power cycle the modem and reconnect to the internet.