IND-EFD: fix to post measurements after failures.
Add timeout parameter to http post.

Allow http post exception to propagate to outer calling function where
it can accumulate the data and repost on the next iteration.

Moved notifying the queue object that items have been processed from the
end of the function to where the objects are stored in the acculation list.