IND-EFD: [#75] Fix fstab entries.
Change BOOT VFAT partition to not auto mount and to have stricter permissions.

Change rootfs device to be generic so that the correct device is determined
by the kernel bootargs.  This caters for the dual boot setup where the rootfs
can be on /dev/mmcblk0p2 or /dev/mmcblk0p3.