IND-EFD: mount /boot/flash read-only at boot.
Mount /dev/mmcblk0p1 to /boot/flash read-only at boot time.
It was set to `noauto` but this leaves the mount point read/writeable
by root and is easy to copy files there assuming the partition is mounted
when infact it is not.

Mounting the partition read-only provides the safety needed, and any copy
of files will fail, which the root user can then mount with
`mount -o remount,rw /boot/flash`