IND-EFD: test_fpga.py, manual triggered sampling.
* Add test_fpga.py test utility to check for the sampling/DMA FPGA system
for buffer overruns, underruns and peak detection errors.

* Modify the ind.py driver interface library to allow software to manually
trigger the sampling process, rather than via the PPS signal.

This allows for 10M samples to be captured and tested with peak-detection,
which is not possible if using PPS as the software peak-detection (used to
compare the hardware peak-detection) is too slow when the data is stored
in non-cahced memory.