[BB#89] Append gps location, battery, temperature to measurements.
SQUAHSED

[BB#89] Append gps location, battery, temperature to measurements.

* Append gps location, battery and solar voltages, and temperature to
the measurements logs and post.

NOTE: battery voltage, solar voltage and temperature are fixed at 0.0
for now, until reading the i2c data is completed.

NOTE: this is likely to be temporary and is being used to test the
new web server being developed by IND.

[BB#89] Bump version to `0.9+BB#89`
[BB#89] Use config setting to append gps data to measurements log.

* Add feature to set configuraiton to append gps location data
(and battery voltage, solar voltage, box temperature), from the
user settings file or the command line.

* Only append extra data to the CSV header and CSV records if the
configuration setting is enabled.

[comments] Change block comment from `##` to `#!`
[eclipse] Change eclipse project name from to `ind-efd-brickstrap`.
[BB#2] Add kernel and device tree with i2c support.

* Device tree supports the LM75 and MCP3021 devices.

[BB#2] Add lm-sensors deb package and efd config

* Add debian `lm-sensors` package.
* Create and `efd.conf` file in `/etc/sensors.d/` to specify devices and
properties.

[config] Add `debian.mirror.serversaustralia.com.au`

[BB#89] Create EFD `Sensors` module to read voltages and battery.

* Create an EFD `Sensors` module, which uses `PySensors` to inteface
  with the lmsensors library.

* Append EFD Sensors data to the measurements logs.

To Do
-----

* Install PySensors via the build system.  PySensors is not a debian
  package so will need to install from sources.  Suggest obtaining
  the source and running `python setup.py install` via the brickstrap
  hooks mechanism.
[PySensors] Add PySensors 0.0.3 source code.

PySensors is a python module to access the `lmsensors` data via
`ctypes`.

PySensors is not available as a standard Debian package, so we need to
install in manually (either from sources or via PyPi using `pip` or
`easyinstall1`)

It was decided to install explicitly from sources as it
(a) means the version will not change if PyPi is updated.
(b) installing `pip` on target installs lots of dependencies (gcc, ...),
    though `easyinstall` shuold work ok.

[PySensors] Add hook to install PySensors module.

Add a brickstrap hook script to install the PySensors module on the
target
filesystem.

[README_sepl.md] Add README file for SEPL stuff.

Conflicts:
	ind-efd/config
	ind-efd/root/boot/uImage
	ind-efd/root/opt/sbin/efd_app.py
	ind-efd/root/opt/sbin/efd_config.py
	ind-efd/root/opt/sbin/efd_gps.py
	ind-efd/root/opt/sbin/efd_measurements.py
	ind-efd/root/opt/sbin/ind.py
	ind-efd/root/opt/sbin/test/settings.py