`eval` lines in package files to allow env vars.
Evaluate each line in a package file so that environment variable
substitution can take place.

This allows for specifying various distributions for packages.
e.g. if using backports => specify a package as:
    gpsd/jessie-backports
or
    gpsd/${SUITE}-backports
or
    gpsd/${BACKPORTS}   -- assuming BACKPORTS="${SUITE}-backports"

This allows to keep the package files the same if changing suites
(e.g. from `jessie` to `stretch`), although it probably highly
likely there would be no need for the backport in that case.