[IND,BB#130] Remove unintended multiple power cycles of modem.
The power-off part of the modem reboot scripts was actually power
cycling the modem multiple times.

Rework the scripts so that only one power-off sequence is called.

Also only one power-on sequence.

If anything fails, then it will exit, and the next cycle of the
`check-internet-connection.sh` script should start the process again.