[IND,BB139] Fix user settings file having wrong default after upgrade.
The upgrade `run` script took the file `/opt/sbin/test/settings` as the
template for the new settings file.  This file was actually used for
testing and had the value of `PEAK_DETECTION_MODE=squared`.

It doesn't seen wise to use the `settings` file in the `test` directory
to be the official template file for new settings.

Created a new file `/opt/sbin/settings_new`, which is now the offical
template file for new settings, which the upgrade `run` script will copy
to the filesystem and modify if it needs to.

NOTE: all default values must be explicitly defined `settings_new`
template file, otherwise existing values may not be copied over.