[IND,BB#142] Fix duplicate data in measurement logs/posts.
The change to sending the `self.measurements` object to the
`cloud_thread`, instead the `csv_data` string, introduced a regression
where duplicate data could be stored in the log files and posted to the
web portal.

The root cause was that `self.measurements` object was being **reused**
for each set of measurements and then sent to the `cloud_thread` via a
queue (but only the reference is sent, not a copy of the object).

When large delays occur in the `cloud_thread` (typically on http posts),
multiple measurement objects are queued, but since it is the same
reference then the only data available is of the last set of
measurements.

The fix is to create a new measurements object for each set of
measurements, instead of **reusing** the one object.