[IND,efd_app] Prevent bogus sanity check error message.
Prevent bogus sanity check error message for current bank timestamp not
being the latest in both banks.

This can happen on the very first capture due to timing of restarting
the EFD service and which bank being captured first (or last).

Ensuring that the `prev_bank` is not equal to the current `bank` should
prevent this bogus error.