[IND][modem-power-enable.sh] Detect NetworkManager connection errors
The `nmncli` command doesn't seem to return an error code if it fails
(at least not when trying to connect to an invalid GMS APN).

It does report an error message (with strings "Error" and "error"), so
we will look for these error keys to ensure the command succeeded or
not.