[ind-efd][BB#159] Implement periodic reboot feature
Periodic reboot can be enabled by adding the following in the
user settings file.

The reboot schedule must comply with `cron` syntax (first 5 fields).
The first 5 fields are: minute, hour, day-of-month, month, day-of-week

If not present, commented out, or set to "", then the default action
is no periodic reboots.

PERIODIC_REBOOT_SCHED="0 2 * * *"

Optionaly specify the reboot type: `reboot` or `shutdown`.
`reboot` does a warm reboot.
`shutdown` does a `poweroff`, and the system will be rebooted by the PMC

NOTE: The reboot type is ignored for now, until we read the
      battery voltage to ensure the system will indeed reboot.

PERIODIC_REBOOT_TYPE="reboot"