[ind-efd][BB#160] Implement a chrony watchdog
The `chrony_watchdog.sh` script is called periodically
(every 10 minutes) via cron.

It checks that the `chronyc sources` output shows the GPS clock is
still sourcing chrony time.

If not, then a reboot will be triggered.

Note: the reboot will not be triggered on the first call to the script,
      only on subsequent calls - to allow chrony to synhronise first.