[ind-efd][hooks] Fix chmod or `*.py` and `*.sh`
For some reason I can't do a `chmod *.py` or `chmod *.sh` in the hook
scripts.  It only works when explicitly naming the files.

For now, just list all the known files and chmod them.

It's not pretty but it will do for now as there is a release pending.