[ind-efd][chrony.service] Fix startup on boot (and in general)
Using `systemctl stop` or `systemctl start` or `systemctl restart`
can cause the service to "hang" or not complete.

Uisng `systemctl kill` to kill gpsd seems to work better.

The move to `systemctl restart` was to make sure the gps nmea logger
would shutdown, but it has been proven that the logger needs to be
written not to use `gpscat` (which talks to the device directly instead
of going throuhg `gpsd`).