[ind-efd][BB#172] Fix incorrect owner IDs after system upgrade
The extraction of the files from the root filesystem archive were being
created with the owner/group IDs of the currently running system,
which can be different from the target upgrade system.

Using the `--numeric-owner` option to `tar` during the extraction
process resolves the issue, so that the files are created with the
same IDs of the files in the archive.